Removes index and adds initial prices
