+++ next-app/pages/index.tsx
@@ -1,72 +1,66 @@
+import { useState, useEffect } from 'react'
+import styles from  "../styles/UserHome.module.css"
 import type { NextPage } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
-import styles from '../styles/Home.module.css'
 
-const Home: NextPage = () => {
+function UserHome(): JSX.Element {
+  const [items, setItems] = useState<null | ItemProps[]>(null)
+
+  const getItems = useEffect(() => {
+    async function getData() {
+      try {
+        const res = await fetch('http://localhost:8080/item-prices')
+        const data = await res.json()
+        setItems(data)
+      } catch (e) {
+        console.log(e)
+      }
+    }
+    getData()
+  }, [])
+
+  console.log(items)
+
+  return (<div className={styles['user-favourites']}>
+    <main>
+      <h1>Items</h1>
+      { items ? items.map(item => (
+        <Item   {...item} key={item.id} />
+      ))
+       :
+      <></>
+      }
+    </main>
+  </div>)
+}
+
+interface ItemProps {
+  name: string,
+  id: number,
+  urlsId: number,
+  prices: [
+    {
+      dateTime: string,
+      id: number,
+      itemId: number,
+      supervalu: number | null
+      tesco: number | null
+      dunnes: number | null
+    }
+  ]
+}
+
+function Item(props: ItemProps): JSX.Element {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+  <div className="item">
+    <h3>{props.name}</h3>
+    <div>
+      <span>{"name of store"}</span>
+      <span>€ {9.6.toString()}</span>
     </div>
+  </div>
   )
 }
 
-export default Home
+export default UserHome
