+++ next-app/styles/UserHome.module.css
@@ -1,0 +1,8 @@
+.price-list {
+  padding: auto 2rem;
+}
+
+.user-favourites {
+  padding-inline: 2rem;
+}
+
